Refactor File structure
