+++ src/App.jsx
@@ -1,12 +1,11 @@
-import React from "react";
 import { Routes, Route } from "react-router-dom";
 import LandingPage from "./pages/LandingPage.jsx";
 import LoginPage from "./pages/LoginPage.jsx";
 import AllEventsPage from "./pages/AllEventsPage.jsx";
 import SignupPage from "./pages/SignupPage.jsx";
-import Dashboard from "./pages/Dashboard.jsx";
-import AdminSignup from "./components/AdminSignup.jsx";
-import { UserExamPanel } from "./components/user-exam-panel.jsx";
+import Dashboard from "./pages/AdminDashboardPage.jsx";
+import AdminSignup from "./components/auth/AdminSignup.jsx";
+import ExamPage from "./pages/ExamPage.jsx";
 
 const App = () => {
   return (
@@ -15,7 +14,7 @@ const App = () => {
       <Route path="/login" element={<LoginPage />} />
       <Route path="/signup" element={<SignupPage />} />
       <Route path="/events" element={<AllEventsPage />} />
-      <Route path="/exams" element={<UserExamPanel />} />
+      <Route path="/exams" element={<ExamPage />} />
 
       {/* Admin Routes */}
       <Route path="/admin/signup" element={<AdminSignup />} />
