+++ src/components/auth/AdminSignup.jsx
@@ -1,4 +1,4 @@
-import React, { useRef, useState, useEffect } from "react";
+import { useRef, useState, useEffect } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import axios from "axios";
 import gsap from "gsap";
@@ -18,9 +18,9 @@ import { Input } from "@/components/ui/input";
 import { Checkbox } from "@/components/ui/checkbox";
 import { Label } from "@/components/ui/label";
 import { cn } from "@/lib/utils";
-import { FloatingParticles } from "./floating-particles";
-import { ThreeDCard } from "./three-d-card";
-import { ProgressBar } from "./progress-bar";
+import { FloatingParticles } from "@/components/floating-particles";
+import { ThreeDCard } from "@/components/three-d-card";
+import { ProgressBar } from "@/components/progress-bar";
 
 gsap.registerPlugin(ScrollTrigger);
