+++ src/pages/ExamPage.jsx
@@ -1,0 +1,13 @@
+import { UserExamPanel } from "@/components/user/user-exam-panel";
+
+const ExamPage = () => {
+  return (
+    <>
+      <main className="flex-1 p-4 sm:p-6">
+        <UserExamPanel />
+      </main>
+    </>
+  );
+};
+
+export default ExamPage;
