+++ src/pages/LandingPage.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import Navbar from "../components/Navbar.jsx";
 import Hero from "../components/Hero.jsx";
 import Footer from "../components/Footer.jsx";
