+++ src/pages/SignupPage.jsx
@@ -1,5 +1,4 @@
-import React from "react";
-import Signup from "../components/Signup";
+import Signup from "../components/auth/Signup.jsx";
 
 const SignupPage = () => {
   return (
